glhf

+++ product.py
@@ -1,13 +1,13 @@
 class Product:
     name: str
     description: str
-    price: int
+    price: float
     quantity: int
 
     def __init__(self, name, description, price, quantity):
         self.name = name
         self.description = description
-        self.price = price
+        self.price = float(price)
         self.quantity = quantity
 
     def __str__(self):
